+++ xcube_geodb/core/geodb.py
@@ -88,12 +88,12 @@ class GeoDBError(ValueError):
 
 
 class EventType:
-
     CREATED = "created"
     DROPPED = "dropped"
     RENAMED = "renamed"
     COPIED = "copied"
     MOVED = "moved"
+    READ = "read"
     PUBLISHED = "published"
     UNPUBLISHED = "unpublished"
     PUBLISHED_GS = "published to geoserver"
@@ -177,11 +177,12 @@ class GeoDBClient(object):
         gs_server_port: Optional[int] = None,
         raise_it: bool = True,
     ):
-
         self._use_auth_cache = True
         self._dotenv_file = dotenv_file
         self._database = None
         self._raise_it = raise_it
+        self._do_log_read = True
+
         # Access token is set here or on request
 
         # defaults
@@ -1046,7 +1047,7 @@ class GeoDBClient(object):
             )
 
             self._log_event(EventType.MOVED, f"collection {old_dn} to {new_dn}")
-            return Message(f"Collection moved from {database} to " f"{new_database}")
+            return Message(f"Collection moved from {database} to {new_database}")
         except GeoDBError as e:
             return self._maybe_raise(e)
 
@@ -1113,10 +1114,8 @@ class GeoDBClient(object):
             self.grant_access_to_collection(
                 collection=collection, usr="public", database=database
             )
-            self._log_event(
-                EventType.PUBLISHED, f"collection " f"{database}_{collection}"
-            )
-            return Message(f"Access granted on {database}_{collection} to " f"public.")
+            self._log_event(EventType.PUBLISHED, f"collection {database}_{collection}")
+            return Message(f"Access granted on {database}_{collection} to public.")
         except GeoDBError as e:
             return self._maybe_raise(e)
 
@@ -1262,7 +1261,7 @@ class GeoDBClient(object):
                 prop_type = properties[prop_name]
                 self._log_event(
                     EventType.PROPERTY_ADDED,
-                    f"{{name: {prop_name}, " f"type: {prop_type}}} to collection {dn}",
+                    f"{{name: {prop_name}, type: {prop_type}}} to collection {dn}",
                 )
             self._refresh_capabilities()
             return Message("Properties added")
@@ -1424,6 +1423,8 @@ class GeoDBClient(object):
 
         """
 
+        self._do_log_read = False
+
         try:
             return self.get_collection(
                 collection="user_databases",
@@ -1448,6 +1449,7 @@ class GeoDBClient(object):
 
         """
 
+        self._do_log_read = False
         try:
             res = self.get_collection(
                 collection="user_databases",
@@ -1601,7 +1603,7 @@ class GeoDBClient(object):
         crs = check_crs(crs)
         if crs and srid and srid != crs:
             raise GeoDBError(
-                f"crs {crs} is not compatible with collection's " f"crs {srid}"
+                f"crs {crs} is not compatible with collection's crs {srid}"
             )
 
         crs = crs or srid
@@ -1648,10 +1650,11 @@ class GeoDBClient(object):
                 except GeoDBError as e:
                     return self._maybe_raise(e)
                 except requests.exceptions.ConnectionError as e:
-                    if "Connection aborted" in str(
-                        e
-                    ) and "LineTooLong('got more than 65536 bytes " "when reading header line')" in str(
-                        e
+                    if (
+                        "Connection aborted" in str(e)
+                        and "LineTooLong('got more than 65536 bytes "
+                        "when reading header line')"
+                        in str(e)
                     ):
                         # ignore this error - the ingestion has worked.
                         # see https://github.com/dcs4cop/xcube-geodb/issues/60
@@ -1660,7 +1663,7 @@ class GeoDBClient(object):
                         raise e
         else:
             self._maybe_raise(
-                GeoDBError(f"Error: Format {type(values)} not " f"supported.")
+                GeoDBError(f"Error: Format {type(values)} not supported.")
             )
 
         msg = f"{total_rows} rows inserted into "
@@ -1773,23 +1776,26 @@ class GeoDBClient(object):
 
             headers = {"Accept": "application/vnd.pgrst.object+json"}
 
+            payload = {
+                "collection": dn,
+                "minx": bbox[0],
+                "miny": bbox[1],
+                "maxx": bbox[2],
+                "maxy": bbox[3],
+                "comparison_mode": comparison_mode,
+                "bbox_crs": bbox_crs,
+                "limit": limit,
+                "where": where,
+                "op": op,
+                "offset": offset,
+            }
             r = self._post(
                 "/rpc/geodb_get_by_bbox",
                 headers=headers,
-                payload={
-                    "collection": dn,
-                    "minx": bbox[0],
-                    "miny": bbox[1],
-                    "maxx": bbox[2],
-                    "maxy": bbox[3],
-                    "comparison_mode": comparison_mode,
-                    "bbox_crs": bbox_crs,
-                    "limit": limit,
-                    "where": where,
-                    "op": op,
-                    "offset": offset,
-                },
+                payload=payload,
             )
+
+            self._maybe_log_read(collection, database, str(payload))
 
             js = r.json()["src"]
             if js:
@@ -2005,6 +2011,8 @@ class GeoDBClient(object):
             else:
                 r = self._get(f"/{dn}")
 
+            self._maybe_log_read(collection, database, actual_query)
+
             js = r.json()
 
             if js:
@@ -2076,6 +2084,8 @@ class GeoDBClient(object):
 
             js = r.json()["src"]
 
+            self._maybe_log_read(collection, database)
+
             if js:
                 srid = self.get_collection_srid(collection, database)
                 return self._df_from_json(js, srid)
@@ -2083,6 +2093,15 @@ class GeoDBClient(object):
                 return DataFrame(columns=["Empty Result"])
         except GeoDBError as e:
             return self._maybe_raise(e, return_df=True)
+
+    def _maybe_log_read(self, collection, database, query=None):
+        if self._do_log_read:
+            message = f"read from collection {database}_{collection}"
+            if query:
+                message += f": {query}"
+            self._log_event(EventType.READ, message)
+        else:
+            self._do_log_read = True
 
     def create_index(self, collection: str, prop: str, database: str = None) -> Message:
         """
@@ -2411,7 +2430,7 @@ class GeoDBClient(object):
         try:
             if self._use_winchester_gs:
                 self._delete(
-                    path=f"/geodb_geoserver/" f"{database}/collections/{collection}"
+                    path=f"/geodb_geoserver/{database}/collections/{collection}"
                 )
             else:
                 self._delete(
@@ -2541,7 +2560,7 @@ class GeoDBClient(object):
 
         if not self._is_owner_of(dn):
             raise GeoDBError(
-                f"User {self.whoami} must be owner of collection " f"{dn} to publish."
+                f"User {self.whoami} must be owner of collection {dn} to publish."
             )
 
         path = "/rpc/geodb_group_publish_collection"
@@ -2583,7 +2602,7 @@ class GeoDBClient(object):
 
         if not self._is_owner_of(dn):
             raise GeoDBError(
-                f"User {self.whoami} must be owner of " f"collection {dn} to unpublish."
+                f"User {self.whoami} must be owner of collection {dn} to unpublish."
             )
 
         path = "/rpc/geodb_group_unpublish_collection"
@@ -2618,8 +2637,7 @@ class GeoDBClient(object):
         dn = f"{database}_dummy"
         if not self._is_owner_of(dn):
             raise GeoDBError(
-                f"User {self.whoami} must be owner of database "
-                f"{database} to publish."
+                f"User {self.whoami} must be owner of database {database} to publish."
             )
 
         path = "/rpc/geodb_group_publish_database"
@@ -2646,8 +2664,7 @@ class GeoDBClient(object):
         dn = f"{database}_dummy"
         if not self._is_owner_of(dn):
             raise GeoDBError(
-                f"User {self.whoami} must be owner of database "
-                f"{database} to unpublish."
+                f"User {self.whoami} must be owner of database {database} to unpublish."
             )
 
         path = "/rpc/geodb_group_unpublish_database"
@@ -2880,6 +2897,7 @@ class GeoDBClient(object):
         """
         database = database or self.database
 
+        self._do_log_read = False
         try:
             self.head_collection(collection, database=database)
         except GeoDBError:
@@ -3002,7 +3020,7 @@ class GeoDBClient(object):
         )
         cursor = conn.cursor()
 
-        with open(f"xcube_geodb/sql/geodb.sql") as sql_file:
+        with open("xcube_geodb/sql/geodb.sql") as sql_file:
             sql_create = sql_file.read()
             cursor.execute(sql_create)